Created BatchForm component and using it in Create Batch
