+++ canopeum_frontend/src/components/analytics/batch-modal/CreateBatchModal.tsx
@@ -1,8 +1,9 @@
-/* eslint-disable max-lines -- disable max-lines */
 import { Dialog, DialogActions, DialogContent, DialogTitle } from '@mui/material'
 import { useCallback, useContext, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
+import type { BatchFormDto } from './BatchForm'
+import BatchForm from '@components/analytics/batch-modal/BatchForm'
 import FertilizersSelector from '@components/analytics/FertilizersSelector'
 import ImageUpload from '@components/analytics/ImageUpload'
 import MulchLayersSelector from '@components/analytics/MulchLayersSelector'
@@ -12,7 +13,6 @@ import { SnackbarContext } from '@components/context/SnackbarContext'
 import useApiClient from '@hooks/ApiClientHook'
 import type { FertilizerType, MulchLayerType, Seeds, SiteSummary, Species, TreeType } from '@services/api'
 import { assetFormatter } from '@utils/assetFormatter'
-import { floorNumberValue } from '@utils/formUtils'
 
 type Props = {
   readonly open: boolean,
@@ -69,16 +69,22 @@ const CreateBatchModal = ({ open, site, handleClose }: Props) => {
   const { getApiClient } = useApiClient()
   const { openAlertSnackbar } = useContext(SnackbarContext)
 
-  const [batch, setBatch] = useState<CreateBatchDto>(defaultCreateBatch)
-  const [batchImageURL, setBatchImageURL] = useState<string>()
-  const [sponsorLogoUrl, setSponsorLogoUrl] = useState<string>()
+  const [batch, setBatch] = useState<BatchFormDto>(defaultCreateBatch)
+
+  const handleBatchChange = (batchFormDto: BatchFormDto) => {
+    console.log('BATCH CHANGE batchFormDto:', batchFormDto)
+
+    setBatch(batchFormDto)
+  }
 
   const handleSubmitBatch = async () => {
     const {
       name,
       size,
       soilCondition,
-      sponsor,
+      sponsorName,
+      sponsorWebsiteUrl,
+      sponsorLogo,
       fertilizers,
       mulchLayers,
       supportedSpecies,
@@ -92,6 +98,10 @@ const CreateBatchModal = ({ open, site, handleClose }: Props) => {
       image,
     } = batch
 
+    const sponsorLogoImage = sponsorLogo
+      ? await assetFormatter(sponsorLogo)
+      : undefined
+
     const batchImage = image
       ? await assetFormatter(image)
       : undefined
@@ -100,7 +110,9 @@ const CreateBatchModal = ({ open, site, handleClose }: Props) => {
       await getApiClient().batchClient.create(
         site.id,
         name,
-        sponsor,
+        sponsorName,
+        sponsorWebsiteUrl,
+        sponsorLogoImage,
         size,
         soilCondition,
         plantCount,
@@ -121,41 +133,25 @@ const CreateBatchModal = ({ open, site, handleClose }: Props) => {
         { severity: 'error' },
       )
 
-      return
+      return false
     }
+
     openAlertSnackbar(
       t('analyticsSite.batch-modal.feedback.create-success'),
     )
     resetBatch()
     handleClose('create')
+
+    return true
   }
 
-  const onImageUpload = (file: File) => {
-    setBatch(value => ({ ...value, image: file }))
-    setBatchImageURL(URL.createObjectURL(file))
-  }
-
-  const onSponsorLogoUpload = (file: File) => {
-    // Update sponsor value here?
-    setBatch(value => ({
-      ...value,
-      sponsor: {
-        name: value.sponsor?.name ?? '',
-        websiteUrl: value.sponsor?.websiteUrl ?? '',
-        logo: file,
-      },
-    }))
-    setSponsorLogoUrl(URL.createObjectURL(file))
+  const resetBatch = () => {
+    console.log('RESET HERE')
   }
 
   const onClose = () => {
     resetBatch()
     handleClose()
-  }
-
-  const resetBatch = () => {
-    setBatch({ ...defaultCreateBatch })
-    setBatchImageURL(undefined)
   }
 
   return (
@@ -167,256 +163,7 @@ const CreateBatchModal = ({ open, site, handleClose }: Props) => {
       </DialogTitle>
 
       <DialogContent className='pb-5'>
-        <form>
-          <div className='d-flex flex-column gap-3'>
-            <div>
-              <label className='form-label' htmlFor='batch-name'>
-                {t('analyticsSite.batch-modal.name-label')}
-              </label>
-              <input
-                className='form-control'
-                id='batch-name'
-                onChange={event => setBatch(value => ({ ...value, name: event.target.value }))}
-                type='text'
-                value={batch.name}
-              />
-            </div>
-
-            <div>
-              <label className='form-label text-capitalize' htmlFor='sponsor-name'>
-                {t('analyticsSite.batch-modal.sponsor-name-label')}
-              </label>
-              <input
-                className='form-control'
-                id='sponsor-name'
-                onChange={event =>
-                  setBatch(value => ({
-                    ...value,
-                    sponsor: value.sponsor
-                      ? { ...value.sponsor }
-                      : undefined,
-                  }))}
-                type='text'
-                value={batch.sponsor?.name}
-              />
-            </div>
-
-            <div>
-              <label className='form-label text-capitalize' htmlFor='sponsor-website-url'>
-                {t('analyticsSite.batch-modal.sponsor-website-url-label')}
-              </label>
-              <input
-                className='form-control'
-                id='sponsor-website-url'
-                onChange={event => setBatch(value => ({ ...value, sponsor: undefined }))}
-                type='text'
-                value={batch.sponsor?.websiteUrl}
-              />
-            </div>
-
-            <div>
-              <label className='form-label text-capitalize' htmlFor='sponsor-logo'>
-                {t('analyticsSite.batch-modal.sponsor-logo-label')}
-              </label>
-              <ImageUpload imageUrl={sponsorLogoUrl} onChange={onSponsorLogoUpload} />
-            </div>
-
-            <div>
-              <label className='form-label text-capitalize' htmlFor='size'>
-                {t('analyticsSite.batch-modal.size-label')}
-              </label>
-              <div className='input-group'>
-                <input
-                  className='form-control'
-                  id='size'
-                  min={0}
-                  onChange={event =>
-                    setBatch(value => ({
-                      ...value,
-                      size: Number.parseInt(event.target.value, 10),
-                    }))}
-                  type='number'
-                  value={floorNumberValue(batch.size)}
-                />
-                <span className='input-group-text'>
-                  {t('analyticsSite.batch-modal.feet-squared')}
-                </span>
-              </div>
-            </div>
-
-            <div>
-              <TreeSpeciesSelector
-                label='analyticsSite.batch-modal.number-of-trees-label'
-                onChange={useCallback(
-                  species =>
-                    setBatch(current => ({
-                      ...current,
-                      species,
-                    })),
-                  [],
-                )}
-                species={batch.species}
-              />
-            </div>
-
-            <div>
-              <label className='form-label' htmlFor='soil-condition'>
-                {t('analyticsSite.batch-modal.soil-condition-label')}
-              </label>
-              <input
-                className='form-control'
-                id='soil-condition'
-                onChange={event =>
-                  setBatch(value => ({ ...value, soilCondition: event.target.value }))}
-                type='text'
-                value={batch.soilCondition}
-              />
-            </div>
-
-            <FertilizersSelector
-              fertilizers={batch.fertilizers}
-              onChange={useCallback(
-                fertilizers =>
-                  setBatch(current => ({
-                    ...current,
-                    fertilizers,
-                  })),
-                [],
-              )}
-            />
-
-            <MulchLayersSelector
-              mulchLayers={batch.mulchLayers}
-              onChange={useCallback(
-                mulchLayers =>
-                  setBatch(current => ({
-                    ...current,
-                    mulchLayers,
-                  })),
-                [],
-              )}
-            />
-
-            <SupportSpeciesSelector
-              onChange={useCallback(
-                supportedSpecies =>
-                  setBatch(current => ({
-                    ...current,
-                    supportedSpecies,
-                  })),
-                [],
-              )}
-              species={batch.supportedSpecies}
-            />
-
-            <div>
-              <label className='form-label text-capitalize' htmlFor='total-number-of-plants'>
-                {t('analyticsSite.batch-modal.total-number-of-plants-label')}
-              </label>
-              <input
-                className='form-control'
-                id='total-number-of-plants'
-                onChange={event =>
-                  setBatch(value => ({
-                    ...value,
-                    plantCount: Number.parseInt(event.target.value, 10),
-                  }))}
-                type='number'
-                value={floorNumberValue(batch.plantCount)}
-              />
-            </div>
-
-            <div>
-              <label className='form-label text-capitalize' htmlFor='survived'>
-                {t('analyticsSite.batch-modal.survived-label')}
-              </label>
-              <input
-                className='form-control'
-                id='survived'
-                onChange={event =>
-                  setBatch(value => ({
-                    ...value,
-                    survivedCount: Number.parseInt(event.target.value, 10),
-                  }))}
-                type='number'
-                value={floorNumberValue(batch.survivedCount)}
-              />
-            </div>
-
-            <div>
-              <label className='form-label text-capitalize' htmlFor='replaced'>
-                {t('analyticsSite.batch-modal.replaced-label')}
-              </label>
-              <input
-                className='form-control'
-                id='replaced'
-                onChange={event =>
-                  setBatch(value => ({
-                    ...value,
-                    replaceCount: Number.parseInt(event.target.value, 10),
-                  }))}
-                type='number'
-                value={floorNumberValue(batch.replaceCount)}
-              />
-            </div>
-
-            <div>
-              <TreeSpeciesSelector
-                label='analyticsSite.batch-modal.seeds-per-species-label'
-                onChange={useCallback(
-                  species =>
-                    setBatch(current => ({
-                      ...current,
-                      seeds: species,
-                    })),
-                  [],
-                )}
-                species={batch.seeds}
-              />
-            </div>
-
-            <div>
-              <label className='form-label text-capitalize' htmlFor='totalNumberSeed'>
-                {t('analyticsSite.batch-modal.total-seeds-label')}
-              </label>
-              <input
-                className='form-control'
-                id='totalNumberSeed'
-                onChange={event =>
-                  setBatch(value => ({
-                    ...value,
-                    totalNumberSeed: Number.parseInt(event.target.value, 10),
-                  }))}
-                type='number'
-                value={floorNumberValue(batch.totalNumberSeed)}
-              />
-            </div>
-
-            <div>
-              <label className='form-label text-capitalize' htmlFor='propagation'>
-                {t('analyticsSite.batch-modal.propagation-label')}
-              </label>
-              <input
-                className='form-control'
-                id='propagation'
-                onChange={event =>
-                  setBatch(value => ({
-                    ...value,
-                    totalPropagation: Number.parseInt(event.target.value, 10),
-                  }))}
-                type='number'
-                value={floorNumberValue(batch.totalPropagation)}
-              />
-            </div>
-
-            <div className='mb-3'>
-              <label className='form-label text-capitalize' htmlFor='batch-image'>
-                {t('analyticsSite.batch-modal.images-label')}
-              </label>
-              <ImageUpload imageUrl={batchImageURL} onChange={onImageUpload} />
-            </div>
-          </div>
-        </form>
+        <BatchForm handleBatchChange={handleBatchChange} />
       </DialogContent>
 
       <DialogActions>
